backend/requirements.txt
Django==3.2.3
djangorestframework==3.15.2
Pillow==9.3.0
djoser==2.2.3